Refactor para reduzir tamanho do main.py; Segmentação de estruturas; Testes de envio com Evolution e autenticação funcionais;
