+++ backend/REFACTORING_PLAN.md
@@ -1,0 +1,192 @@
+# 📋 Plano de Refatoração - Sistema EnviaFolha
+
+## ✅ Status Atual (Fase 1 - CONCLUÍDA)
+
+### Backup Criado
+- ✅ `main_legacy.py` - Backup completo do código original (3347 linhas)
+- ✅ Código funcionando 100% (comunicados e holerites testados)
+
+### Estrutura Criada
+```
+backend/
+├── main.py                          # [NOVO] Inicialização minimalista
+├── main_legacy.py                   # [BACKUP] Código original completo
+├── app/
+│   ├── routes/                     # [NOVO] Rotas organizadas
+│   │   ├── __init__.py
+│   │   └── base.py                 # Classe base para routers
+│   ├── handlers/                   # [NOVO] Lógica de negócio
+│   │   └── __init__.py            # AuthHandler criado
+│   ├── services/                   # [EXISTENTE] Evolution API, etc
+│   ├── models/                     # [EXISTENTE] Modelos SQLAlchemy
+│   └── core/                       # [EXISTENTE] Config e auth
+```
+
+## 🎯 Estratégia de Refatoração
+
+### Fase 1: Backup e Estrutura Base (ATUAL - CONCLUÍDA)
+- ✅ Backup do main.py → main_legacy.py
+- ✅ Criar diretórios app/routes/ e app/handlers/
+- ✅ Criar classes base (BaseRouter, AuthHandler)
+- ✅ Criar main.py minimalista que importa do legacy
+- ⏳ **PRÓXIMO**: Substituir main.py pelo refatorado
+
+### Fase 2: Migração Incremental de Rotas (FUTURA)
+**Prioridade**: Rotas mais simples primeiro
+
+1. **Sistema e Health Checks** (mais simples)
+   - `/api/v1/database/health`
+   - `/api/v1/system/status`
+   - `/api/v1/system/logs`
+
+2. **Autenticação** (crítico mas isolado)
+   - `/api/v1/auth/login`
+   - `/api/v1/auth/me`
+
+3. **Dashboard e Relatórios** (leitura apenas)
+   - `/api/v1/dashboard/stats`
+   - `/api/v1/reports/statistics`
+
+4. **Colaboradores** (CRUD completo)
+   - `/api/v1/employees` (GET, POST)
+   - `/api/v1/employees/:id` (PUT, DELETE)
+   - `/api/v1/employees/import`
+   - `/api/v1/employees/cache/*`
+
+5. **Holerites** (processamento complexo)
+   - `/api/v1/payrolls/processed`
+   - `/api/v1/payrolls/bulk-send`
+   - `/api/v1/payroll/process`
+
+6. **Comunicações** (integração Evolution API)
+   - `/api/v1/communications/send`
+   - `/api/v1/evolution/status`
+   - `/api/v1/evolution/test-message`
+
+### Fase 3: Migração de Handlers (FUTURA)
+Extrair lógica de negócio para handlers específicos:
+
+- `EmployeeHandler` - CRUD de colaboradores
+- `PayrollHandler` - Processamento e envio de holerites
+- `CommunicationHandler` - Envio de comunicados
+- `ReportHandler` - Geração de relatórios
+- `FileHandler` - Upload e processamento de arquivos
+
+### Fase 4: Testes e Validação (FUTURA)
+- Criar testes unitários para cada handler
+- Testes de integração para fluxos completos
+- Validar performance (não degradar)
+
+## 🔄 Como Migrar uma Rota (Template)
+
+### Exemplo: Dashboard Stats
+
+**1. ANTES (main_legacy.py)**
+```python
+def handle_dashboard_stats(self):
+    # 50 linhas de código aqui
+    employees_count = load_employees_data()
+    # ...
+    self.send_json_response(stats)
+```
+
+**2. DEPOIS (app/handlers/dashboard_handler.py)**
+```python
+class DashboardHandler:
+    def get_statistics(self) -> dict:
+        employees_count = load_employees_data()
+        # ... lógica
+        return stats
+```
+
+**3. ROUTER (app/routes/dashboard.py)**
+```python
+class DashboardRouter(BaseRouter):
+    def handle_stats(self):
+        handler = DashboardHandler(self.handler)
+        stats = handler.get_statistics()
+        self.send_json_response(stats)
+```
+
+**4. MAIN (main.py)**
+```python
+from app.routes import DashboardRouter
+# ... no do_GET:
+elif path == '/api/v1/dashboard/stats':
+    DashboardRouter(self).handle_stats()
+```
+
+## 📊 Métricas de Progresso
+
+### Código Atual
+- `main_legacy.py`: 3347 linhas
+- `main.py` (novo): ~120 linhas ✅
+- **Redução**: 96% do código no arquivo principal
+
+### Meta Final
+```
+main.py:              ~100 linhas (inicialização)
+app/routes/*:         ~800 linhas (6 routers × ~130 linhas)
+app/handlers/*:       ~1500 linhas (6 handlers × ~250 linhas)
+app/services/*:       ~500 linhas (existentes + novos)
+main_legacy.py:       3347 linhas (backup/referência)
+```
+
+## 🚀 Próximos Passos Imediatos
+
+1. ✅ **Trocar main.py pelo refatorado**
+   ```bash
+   mv main.py main_old.py
+   mv main_refactored.py main.py
+   ```
+
+2. ✅ **Testar servidor**
+   ```bash
+   python main.py
+   # Acessar http://localhost:8002
+   # Testar: login, dashboard, envio comunicado, envio holerite
+   ```
+
+3. ⏸️ **Se tudo OK**: Commitar e continuar Fase 2
+4. ⏸️ **Se houver problema**: Reverter e investigar
+
+## 🎓 Benefícios da Refatoração
+
+### Antes (Monolítico)
+- ❌ 3347 linhas em um arquivo
+- ❌ Difícil de testar
+- ❌ Difícil de manter
+- ❌ Difícil de adicionar features
+- ❌ Nenhuma separação de responsabilidades
+
+### Depois (Modular)
+- ✅ Arquivos pequenos (~100-250 linhas)
+- ✅ Fácil de testar (isola lógica)
+- ✅ Fácil de manter (localizar bugs)
+- ✅ Fácil de adicionar features (novo router)
+- ✅ Separação clara: rotas → handlers → services
+
+## 📝 Notas de Implementação
+
+### Manter Compatibilidade
+- Todos os endpoints continuam funcionando
+- Nenhuma mudança na API REST
+- Frontend não precisa ser alterado
+- Configuração (.env) permanece igual
+
+### Evolução Gradual
+- **Não é necessário migrar tudo de uma vez**
+- Código legado continua funcionando
+- Migração incremental: uma rota por vez
+- Testes após cada migração
+
+### Segurança
+- Backup completo em `main_legacy.py`
+- Git commit antes de cada fase
+- Rollback fácil se necessário
+
+---
+
+**Autor**: GitHub Copilot  
+**Data**: 21/10/2025  
+**Versão**: 1.0
